Successfully reading files in order Asynchronously
Reading the files in, storing each as an object (which contains their filename, contents, the order number of when they were requested, and the order in which they returned their results) in a global array. Once all files have been read in, sorting the global array based on the order number contained in each one of it's objects, and then finally logging out the details of the file in the order they were requested (regardless of when their data was returned.