Successfully Watching my Directory
Successfully using fs.watch() to alert me of changes in my directory. Every time a file is changed, I am using the fs.mkdir() method to create a new folder with the current epoch time in the title, and storing a copy of the new file in there. This file was originally called app-copyAllFiles-addLines-newDir but I have now added 'watch' to the title (app-copyAllFiles-addLines-watch-newDir)