Creating global ignoreFiles
Creating a ignoreFiles.txt file, which contains a list of all the files I want to ignore. Reading this in at the start of every app (using the fs.readFileSync() function to ensure that the ignoreFiles array is initialised before anything else takes place). This way I now only have to update that file, instead of reconfiguring every one of the apps when I add a new folder, for example.